+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/DriveRR/opmode/ManualFeedforwardTuner.java
@@ -1,11 +1,11 @@
-package org.firstinspires.ftc.teamcode.drive.opmode;
+package org.firstinspires.ftc.teamcode.DriveRR.opmode;
 
-import static org.firstinspires.ftc.teamcode.drive.DriveConstants.MAX_ACCEL;
-import static org.firstinspires.ftc.teamcode.drive.DriveConstants.MAX_VEL;
-import static org.firstinspires.ftc.teamcode.drive.DriveConstants.RUN_USING_ENCODER;
-import static org.firstinspires.ftc.teamcode.drive.DriveConstants.kA;
-import static org.firstinspires.ftc.teamcode.drive.DriveConstants.kStatic;
-import static org.firstinspires.ftc.teamcode.drive.DriveConstants.kV;
+import static org.firstinspires.ftc.teamcode.DriveRR.DriveConstants.MAX_ACCEL;
+import static org.firstinspires.ftc.teamcode.DriveRR.DriveConstants.MAX_VEL;
+import static org.firstinspires.ftc.teamcode.DriveRR.DriveConstants.RUN_USING_ENCODER;
+import static org.firstinspires.ftc.teamcode.DriveRR.DriveConstants.kA;
+import static org.firstinspires.ftc.teamcode.DriveRR.DriveConstants.kStatic;
+import static org.firstinspires.ftc.teamcode.DriveRR.DriveConstants.kV;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
@@ -22,7 +22,7 @@ import com.qualcomm.robotcore.hardware.VoltageSensor;
 import com.qualcomm.robotcore.util.RobotLog;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
-import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
+import org.firstinspires.ftc.teamcode.DriveRR.SampleMecanumDrive;
 
 import java.util.Objects;
