+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Toros/Opmode.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Toros;
 //Package means where the class is stored
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.Servo;
