+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Toros/Swaws.java
@@ -1,4 +1,4 @@
-    package org.firstinspires.ftc.teamcode;
+    package org.firstinspires.ftc.teamcode.Toros;
 
     import com.qualcomm.robotcore.eventloop.opmode.Disabled;
     import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
